+++ hipamd/src/hip_platform.hpp
@@ -82,6 +82,8 @@ class PlatformState {
   std::shared_ptr<UniqueFD> GetUniqueFileHandle(const std::string& file_path);
   bool CloseUniqueFileHandle(const std::shared_ptr<UniqueFD>& ufd);
 
+  void GetLoadingMode(hipModuleLoadingMode_t* mode);
+
   // Logging lock accessor
   amd::Monitor& GetLogLock() { return lg_lock_; }