+++ rocclr/utils/flags.hpp
@@ -277,7 +277,10 @@ release(uint, HIP_SKIP_ABORT_ON_GPU_ERROR, true,                              \
 release(bool, HIP_FORCE_SPIRV_CODEOBJECT, false,                              \
         "Force use of SPIRV instead of device specific code object.")         \
 release(uint, DEBUG_CLR_BATCH_CPU_SYNC_SIZE, 8,                               \
-        "Forces the minimum batch size for CPU sync")  // clang-format on
+        "Forces the minimum batch size for CPU sync")                         \
+release(cstring, HIP_MODULE_LOADING , "lazy",                                 \
+        "Defers module loading to a later point. Default is lazy")            \
+
 
 namespace amd {